+++ packages/svelte/jest.config.js
@@ -1,7 +1,8 @@
 module.exports = {
-    transform: {
-        '^.+\\.jsx?$': 'babel-jest'
-    },
-    testRegex: '(/__tests__/.*|(\\.|/)(test|spec))\\.js?$',
-    moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node']
-}
+  transform: {
+    '^.+\\.jsx?$': 'babel-jest'
+  },
+  transformIgnorePatterns: ['node_modules/(?!@easyroute/core)'],
+  testRegex: '(/__tests__/.*|(\\.|/)(test|spec))\\.js?$',
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node']
+};